Mudança na formula de calculo na coluna Qtd Emb

+++ src/main/kotlin/br/com/astrosoft/produto/model/beans/ProdutoEstoque.kt
@@ -14,7 +14,7 @@ class ProdutoEstoque(
   var unidade: String?,
   var grade: String?,
   var embalagem: Int?,
-  var qtdEmbalagem: Int?,
+  var qtdEmbalagem: Double?,
   var estoque: Int?,
   var locSaci: String?,
   var locApp: String?,
